chore: refactored code and added unit tests
