+++ src/test/kotlin/net/atos/zac/app/informatieobjecten/EnkelvoudigInformatieObjectRestServiceTest.kt
@@ -27,7 +27,7 @@ import net.atos.client.zgw.zrc.model.createZaakInformatieobject
 import net.atos.client.zgw.ztc.ZtcClientService
 import net.atos.client.zgw.ztc.model.createBesluitType
 import net.atos.client.zgw.ztc.model.createInformatieObjectType
-import net.atos.zac.app.audit.converter.RESTHistorieRegelConverter
+import net.atos.zac.app.audit.converter.ZaakHistoryLineConverter
 import net.atos.zac.app.informatieobjecten.converter.RestInformatieobjectConverter
 import net.atos.zac.app.informatieobjecten.converter.RestInformatieobjecttypeConverter
 import net.atos.zac.app.informatieobjecten.converter.RestZaakInformatieobjectConverter
@@ -64,7 +64,7 @@ class EnkelvoudigInformatieObjectRestServiceTest : BehaviorSpec({
     val ontkoppeldeDocumentenService = mockk<OntkoppeldeDocumentenService>()
     val policyService = mockk<PolicyService>()
     val restGerelateerdeZaakConverter = mockk<RestGerelateerdeZaakConverter>()
-    val restHistorieRegelConverter = mockk<RESTHistorieRegelConverter>()
+    val zaakHistoryLineConverter = mockk<ZaakHistoryLineConverter>()
     val restInformatieobjectConverter = mockk<RestInformatieobjectConverter>()
     val restInformatieobjecttypeConverter = mockk<RestInformatieobjecttypeConverter>()
     val webdavHelper = mockk<WebdavHelper>()
@@ -84,7 +84,7 @@ class EnkelvoudigInformatieObjectRestServiceTest : BehaviorSpec({
         zaakInformatieobjectConverter = zaakInformatieobjectConverter,
         restInformatieobjectConverter = restInformatieobjectConverter,
         restInformatieobjecttypeConverter = restInformatieobjecttypeConverter,
-        restHistorieRegelConverter = restHistorieRegelConverter,
+        zaakHistoryLineConverter = zaakHistoryLineConverter,
         restGerelateerdeZaakConverter = restGerelateerdeZaakConverter,
         loggedInUserInstance = loggedInUserInstance,
         webdavHelper = webdavHelper,
